move defaultProps to compat
