+++ src/clone-element.js
@@ -16,18 +16,10 @@ export function cloneElement(vnode, props, children) {
 		ref,
 		i;
 
-	let defaultProps;
-
-	if (vnode.type && vnode.type.defaultProps) {
-		defaultProps = vnode.type.defaultProps;
-	}
-
 	for (i in props) {
 		if (i == 'key') key = props[i];
 		else if (i == 'ref') ref = props[i];
-		else if (props[i] === undefined && defaultProps !== undefined) {
-			normalizedProps[i] = defaultProps[i];
-		} else {
+		else {
 			normalizedProps[i] = props[i];
 		}
 	}
